reduced presentation to fit 20 minutes; updated but not really tested build script

+++ buildScript.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 #this script is not yet tested, just a brainstorming yet...
 
-sudo aptitude install build-essential cmake git libboost-dev libfreeimage-dev libtinyxml-dev libxrandr-dev
+
+sudo aptitude install build-essential cmake git libboost-dev libfreeimage-dev libtinyxml-dev libxrandr-dev libgrantlee-dev
 
 #install boost
 #sudo aptitude install libboost-all-dev
@@ -12,21 +13,41 @@ sudo aptitude install build-essential cmake git libboost-dev libfreeimage-dev li
 
 
 #create a thirdParty directory if it doesn't exist
-cd ..
-if [ ! -d thirdParty ] 
+if [ $# -eq 0 ]
 then
-	mkdir thirdParty
+	echo "no param for third party library directory provided, checking for default location (../thirdparty)..."
+	thirdPartyLibDir="../thirdParty"
+else
+	echo "using $1 as third party library directory"
+	thirdPartyLibDir=$1
 fi
-cd thirdParty
 
 
-#have to get glfw manually, as the git repo has no tags identifying the 2.7 release, and the direct download via wget both produces a strange filename and th 
-#download location does't seem quit permanent
+if [ -d  $thirdPartyLibDir ] 
+then
+	echo "third party library  directory exists ($thirdPartyLibDir), using it"
+else
+	echo "third party library  directory ($thirdPartyLibDir) doesn't exist, creating it"
+	mkdir $thirdPartyLibDir
+fi
+
+
+#save working dir to return to it after completing dependency download
+flewnitDir=`pwd`
+echo "directory of Flewnit library: $flewnitDir;"
+
+#setup the dependencies ----------------------------------------------------------------------------------------
+
+cd $thirdPartyLibDir
+
+#have to get glfw manually, as the git repo has no tags identifying the 2.7 release, 
+#and the direct download via wget both produces a strange filename and the 
+#download location does't seem quite permanent
 if [ ! -d glfw-2.7 ] 
 then
-	echo you have to download GLFW 2.7 and unpack it into thirdparty/glfw-2.7; press enter when you\'re done
+	echo "you have to download GLFW 2.7 manually and unpack it into $thirdPartyLibDir/glfw-2.7\; press enter when you\'re done"
 	read
-else
+	#compile and install
 	cd glfw-2.7
 	sudo make x11-install
 	cd ..
@@ -34,9 +55,12 @@ fi
 
 
 #get glm math library from git
-if [ ! -d ogl-math ] 
+#if [ ! -d glm ] 
+if [ ! -d  ogl-math]
 then
 	git clone git://ogl-math.git.sourceforge.net/gitroot/ogl-math/ogl-math
+	#rename folder
+	#mv ogl-math glm
 fi
 
 
@@ -58,17 +82,30 @@ then
 	#wait for user	
 	read
 fi
-	
+#compile and install assimp
 make -j9
 sudo make install
 cd ..
 
-#return to original Flewnit dir:
-cd ../Flewnit
+
+#----------------------------------------------------------------------------------------
+
+#all dependencies should have been setup, so let's build Flewnit, finally
+cd $flewnitDir
+	
 make clean
-rm CMakeCache.txt
-rm -r CMakeFiles
-cmake .
+
+echo "do you want to delete the cmake cache? y/n"
+read answer
+if [ "$answer" == "y" ]
+then
+	echo "removing cache"
+	rm CMakeCache.txt
+	rm -r CMakeFiles
+fi
+
+
+cmake-gui .
 make -j9
